requirements.txt
streamlit
antlr4-python3-runtime
google-generativeai